+++ certbot/certbot/tests/util.py
@@ -126,7 +126,12 @@ def load_comparable_csr(*names: str) -> jose.ComparableX509:
     return jose.ComparableX509(load_csr(*names))
 
 
-def load_rsa_private_key(*names: str) -> jose.ComparableRSAKey:
+def load_jose_rsa_private_key_pem(*names: str) -> jose.ComparableRSAKey:
+    """Load RSA private key wrapped in jose.ComparableRSAKey"""
+    return jose.ComparableRSAKey(load_rsa_private_key_pem(*names))
+
+
+def load_rsa_private_key_pem(*names: str) -> RSAPrivateKey:
     """Load RSA private key."""
     loader = _guess_loader(names[-1], crypto.FILETYPE_PEM, crypto.FILETYPE_ASN1)
     loader_fn: Callable[..., Any]
@@ -134,16 +139,9 @@ def load_rsa_private_key(*names: str) -> jose.ComparableRSAKey:
         loader_fn = serialization.load_pem_private_key
     else:
         loader_fn = serialization.load_der_private_key
-    return jose.ComparableRSAKey(
-        cast(RSAPrivateKey,
-             loader_fn(load_vector(*names), password=None, backend=default_backend())))
-
-
-def load_pyopenssl_private_key(*names: str) -> crypto.PKey:
-    """Load pyOpenSSL private key."""
-    loader = _guess_loader(
-        names[-1], crypto.FILETYPE_PEM, crypto.FILETYPE_ASN1)
-    return crypto.load_privatekey(loader, load_vector(*names))
+    key = loader_fn(load_vector(*names), password=None, backend=default_backend())
+    assert isinstance(key, RSAPrivateKey)
+    return key
 
 
 def make_lineage(config_dir: str, testfile: str, ec: bool = True) -> str:
@@ -449,7 +447,7 @@ def lock_and_call(callback: Callable[[], Any], path_to_lock: str) -> None:
     process.start()
 
     # Wait confirmation that lock is acquired
-    assert receive_event.wait(timeout=10), 'Timeout while waiting to acquire the lock.'
+    assert receive_event.wait(timeout=20), 'Timeout while waiting to acquire the lock.'
     # Execute the callback
     callback()
     # Trigger unlock from foreign process